Move charsToString to architecture dependent source to fix casting problem
