+++ pkg/platform/architecture_linux.go
@@ -14,15 +14,3 @@ func GetRuntimeArchitecture() (string, error) {
 	}
 	return charsToString(utsname.Machine), nil
 }
-
-func charsToString(ca [65]int8) string {
-	s := make([]byte, len(ca))
-	var lens int
-	for ; lens < len(ca); lens++ {
-		if ca[lens] == 0 {
-			break
-		}
-		s[lens] = uint8(ca[lens])
-	}
-	return string(s[0:lens])
-}
